feat: 중복 확인 및 닉네임 검증 유틸리티 분리
- duplicateUtils.js 생성으로 모든 검증 로직 재사용 가능
- 중복 확인, 닉네임 검증, 검증 메시지 관리 통합
- 회원가입에서 분리된 유틸리티 적용 완료
- 닉네임 입력 시 주소 초기화 버그 수정
